+++ Week1/Online Lecture/IndexInSequence.py
